jwt authentication

+++ Backend/controller/labour_controller.js
@@ -1,9 +1,10 @@
 import { labourModel } from "../Model/constructionModel.js"
 
 export const addlabour=async (req,res)=>{
+    const userId=req.user.id;
     try {
         const {id,name,age,salary,role,mobile,gender,address,image}=await req.body
-        const labourdata= await labourModel.create({id,name,age,salary,role,mobile,gender,address,image})
+        const labourdata= await labourModel.create({id,name,age,salary,role,mobile,gender,address,image,userId})
         res.status(200).json({message:"add labourdata successfully",labourdata})
     } catch (error) {
         console.log("something error in addlabour :",error.message)
@@ -13,7 +14,8 @@ export const addlabour=async (req,res)=>{
 
 export const getlabour=async (req,res)=>{
     try {
-         const labourdata= await labourModel.find();
+         const userId=req.user.id;
+         const labourdata= await labourModel.find({userId});
          res.status(201).json({success:true,data:labourdata})
     } catch (error) {
         console.log("something error in getlabour :",error.message)
@@ -22,9 +24,10 @@ export const getlabour=async (req,res)=>{
 
 export const editlabour= async(req,res)=>{
     try {
+         const userId=req.user.id;
          const {id,name,age,mobile,gender,address,role,image}=req.body
         
-        const updatelabour=await labourModel.updateOne({id},{$set:req.body});
+        const updatelabour=await labourModel.updateOne({id,userId},{$set:req.body});
         res.status(200).json({msg:'updated successflly!',updatelabour})
         console.log("success ")
     } catch (error) {
@@ -36,7 +39,8 @@ export const editlabour= async(req,res)=>{
 export const deletelabour=async(req,res)=>{
     const deletedata=req.params.id
     try {
-        const deletelabour=await labourModel.deleteOne({id:deletedata})
+        const userId=req.user.id;
+        const deletelabour=await labourModel.deleteOne({id:deletedata,userId})
         console.log(deletelabour)
         res.status(200).json({msg:'successfully deleted'})
     } catch (error) {
